Cleaned up code and deleted header files (they don't work)

+++ uchar_array.c
@@ -1,6 +1,3 @@
-//#include "int_array.h"
-#include <math.h>
-
 typedef struct {
   unsigned char *array;
   double standard_dev;
@@ -9,7 +6,7 @@ typedef struct {
 
 unsigned char max(int_array_t *data) {
   unsigned char result = 0;
-  for (uint8_t i = 0; i < data->size; i++) {
+  for (int i = 0; i < data->size; i++) {
     if (data->array[i] > result) {
       result = data->array[i];
     }
@@ -29,9 +26,11 @@ void standard_dev(int_array_t *data) {
   double average = avg(data);
   double total = 0;
   printf("%i\n", data->size);
+
   for (int i = 0; i < data->size; i++) {
     total += pow(data->array[i] - average, 2);
   }
+
   total /= data->size;
   data->standard_dev = sqrt(total);
 }
